Added calc construction and get functions for I Format

+++ RFormat.h
@@ -20,7 +20,7 @@ class RFormat:public InstructionSet{
 
 public:
     // --- Constructor ---
-    RFormat(int hexInstruction, unsigned int address);
+    RFormat(int64_t hexInstruction, unsigned int address);
 
     // --- Calc Instructions Funcs ---
     void calcOPCode();
@@ -33,7 +33,8 @@ public:
 //    void setMemory1(unsigned int mem1);
 //    void setMemory2(unsigned int mem2);
 //    void setDestReg(unsigned int destReg);
-//    void setFuncCode(unsigned int funcCode);
+//    void setFuncCode(unsigned int funcCode);+
+
 
     // --- Get Functions ---
 public:
